TASK-1: As a user I can't add a todo without a description.
As a part of this task, a test suite, using unittest, was created for the project. Validations for required fields were added to SQLAlchemy models and tests were written to check integrity of the records written to the database. Also, I wrote some tests to check server response for correct and incorrect user and todo creation, as well as user login/logout functionalities. I will continue adding tests to the test suite in subsequent tasks.
Modified layout and controller to set flash message according to action and with appropriate color.
Removed some redundant routes with no trailing slash from controller.