EXTRA: Hiding and displaying completed todos
Users should have a way to hide or display completed todos at will.
Implemented a way to accomplish that by setting a cookie in the browser to store that preference.
Analyzed the option of storing the configuration in the database, but since losing the preference is not a big deal I really didn't feel that storing that kind of unimportant information in the database was the way to go.